PTSL_beattool/requirements.txt
py-ptsl>=0.5.0